moved parseExpression into a new file called expression.cpp
