+++ lib/parser/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library(Parser STATIC parser.cpp syntax/function.cpp syntax/type.cpp syntax/return.cpp syntax/value.cpp)
+add_library(Parser STATIC parser.cpp syntax/function.cpp syntax/type.cpp syntax/return.cpp syntax/value.cpp syntax/expression.cpp)
